# svc/catalog/openapi.yaml
openapi: "3.0.0"
info:
  version: 1.0.0
  title: Catalog
  description: Service responsible for managing catalog items
servers:
  - url: /
paths:
  /api-docs:
    get:
      summary: Swagger documentation
      description: Returns Swagger documentation
      responses:
        200:
          description: Swagger documentation
        500:
          description: Error response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /healtz:
    get:
      summary: Health endpoint
      responses:
        200:
          description: Service is healthy

  /api/v1/items:
    get:
      summary: Returns all items
      operationId: getItems
      description: Returns an items from the catalog.
      parameters:
        - name: pageSize
          in: query
          description: Number of elements to be returned. Default 100
          schema:
            type: integer
            default: 100
            minimum: 1
        - name: page
          in: query
          description: Page number.
          schema:
            type: integer
            default: 1
            minimum: 1
      responses:
        200:
          description: Items response
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ItemResponse'
    post:
      summary: Create new item
      operationId: createItem
      description: Creates an item in catalog.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/NewItemRequest'
      responses:
        201:
          description: Item response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ItemResponse'
        500:
          description: Error response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/v1/items/{id}:
    get:
      summary: Returns an item by ID
      operationId: findItemByID
      description: Returns an item from the catalog.
      parameters:
        - name: id
          in: path
          description: ID of an item to fetch
          required: true
          schema:
            type: integer
            format: uint
      responses:
        200:
          description: Item response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ItemResponse'
        500:
          description: Error response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    delete:
      summary: Removes an item by ID
      operationId: deleteItemByID
      description: Deletes item from the catalog based on the ID supplied
      parameters:
        - name: id
          in: path
          description: ID of an item to delete
          required: true
          schema:
            type: integer
            format: uint
      responses:
        200:
          description: Item deleted
        500:
          description: Error response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    put:
      summary: Updates an item by ID
      operationId: updateItemByID
      description: Updates an item in catalog based on the ID supplied
      parameters:
        - name: id
          in: path
          description: ID of an item to update
          required: true
          schema:
            type: integer
            format: uint
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateItemRequest'
      responses:
        200:
          description: Item updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ItemResponse'
        500:
          description: Error response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
components:
  schemas:
    NewItemRequest:
      required:
        - name
        - description
        - price
        - priceCode
      properties:
        name:
          type: string
          description: Name of the item
        description:
          type: string
          description: Description of the item
        price:
          type: number
          format: double
          description: Price of the item
        priceCode:
          type: string
          description: Currency of the price
    UpdateItemRequest:
      properties:
        name:
          type: string
          description: Name of the item
        description:
          type: string
          description: Description of the item
        price:
          type: number
          format: double
          description: Price of the item
        priceCode:
          type: string
          description: Currency of the price
    ItemResponse:
      properties:
        id:
          type: integer
          format: uint
          description: Unique ID of the item
        name:
          type: string
          description: Name of the item
        description:
          type: string
          description: Description of the item
        price:
          type: number
          format: double
          description: Price of the item
        priceCode:
          type: string
          description: Currency of the price
    ErrorResponse:
      required:
        - message
      properties:
        message:
          type: string
          description: Error message